+++ comic_dl/sites/gomanga.py
@@ -71,7 +71,7 @@ def single_chapter(url,current_directory, logger):
     print('\n')
     print("Completed downloading ",Series_Name)
 
-def whole_series(url,current_directory, logger):
+def whole_series(url,current_directory, logger, sortingOrder):
     if not url:
         print("Couldn't get the URL. Please report it on Github Repository.")
     
@@ -91,15 +91,28 @@ def whole_series(url,current_directory, logger):
     soup = BeautifulSoup(Page_source, 'html.parser')
 
     chapter_text = soup.findAll('div',{'class':'title'})
+    all_links = []
+
     
     for link in chapter_text:
         x = link.findAll('a')
         for a in x:
             url = a['href']
             debug("Final URL : %s" % url)
-            single_chapter(url,current_directory, logger)
+            all_links.append(url)
 
-def gomanga_Url_Check(input_url,current_directory, logger):
+    # print(all_links)
+
+    if str(sortingOrder).lower() in ['new','desc','descending','latest']:
+        for chapLink in all_links:
+            single_chapter(chapLink, current_directory, logger)
+    elif str(sortingOrder).lower() in ['old','asc','ascending','oldest']:
+        # print("Running this")
+        for chapLink in all_links[::-1]:
+            single_chapter(chapLink, current_directory, logger)
+    print("Finished Downloading")
+
+def gomanga_Url_Check(input_url,current_directory, logger, sortingOrder):
     if logger == "True":
         basicConfig(format='%(levelname)s: %(message)s', filename="Error Log.log", level=DEBUG)
     
@@ -124,7 +137,7 @@ def gomanga_Url_Check(input_url,current_directory, logger):
             match = found.groupdict()
             if match['comic']:
                 url = str(input_url)
-                whole_series(url,current_directory, logger)
+                whole_series(url,current_directory, logger, sortingOrder)
             else:
                 pass